builtin-run: Allow octal and hex numbers for -m/--mem
There is no reason why -m/--mem should not allow the user to express the
desired memory size in octal or hexadecimal, especially when it's as easy
as changing the strtoull() 'base' parameter to 0, as per man 3 strtoull.

Before:

  $ ./lkvm run -m 0x200 -k Image

would fail with the error message:

  Fatal: Invalid RAM size: 0x200

And now that works as expected.

Before:

  $ ./lkvm run -m 01000 -k Image

would create a VM with 1000MB of memory, when it's known that numbers
that start with a 0 are in base 8. With this patch, that's interpreted
correctly as 512 in base 8.

Note that this is a change in behaviour, but writing decimal numbers with
leading zeros is very uncommon, and it is this author's humble opinion that
there are no kvmtool users that do this.

Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>
Link: https://lore.kernel.org/r/20241128135041.8737-1-alexandru.elisei@arm.com
Signed-off-by: Will Deacon <will@kernel.org>