Propagate the error code from any VCPU thread
kvm_cmd_run_work() doesn't capture the return code for VCPU 0 when calling
pthread_join(). Then kvm_cpu__exit() waits for the threads for VCPUs 1..N
to terminate, and it sets the return value to 0 if at least one of the VCPU
threads returns 0.

This approach creates several issues:

1. An error code from a previous VCPU is overwritten with 0 (success).
2. kvmtool will return 0 even if VCPU 0 encountered an error, as long as
   at least one of the other VCPUs exited successfully.
3. kvm_cpu__exit() will return an uninitialized value if all VCPUs 1..N
   exited with an error, or if there is only one VCPU, VCPU 0.

Fix all issues issues by propagating the return code from VCPU 0, and
saving the error code for each of the remaining VCPUs. If multiple VCPU
threads exit with an error, the exit status is set to the first error.

With this change, kvmtool will terminate with a non-zero exit code if there
was an unhandled error in KVM_RUN, as expected.

While on the subject of return values, apply the more common pattern of
using PTR_ERR(<return value>) in kvm_cpu_thread() instead of double
casting the return value.

Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>
Link: https://lore.kernel.org/r/20250120161800.30270-2-alexandru.elisei@arm.com
Signed-off-by: Will Deacon <will@kernel.org>