Do not a print a warning on failing host<->guest address translation
guest_flat_to_host() and host_to_guest_flat return NULL, respectively 0,
when the address is not found in the existing memslots. It is expected that
the calling code will handle this error.

However, both functions also print an error message containing the
offending address. This can be redundant, if the calling code also prints
the address, or even misleading, if the calling code can gracefully handle
the failure, like is the case in kvm__dump_mem().

Change the warning to a debug, since knowing the address might still be
useful for those call sites where the address isn't printed, or if the
error isn't handled at all.

Before, when running the PMU test from kvm-unit-tests using the test
runner, which redirects stdout (where kvm__dump_mem() writes) and stderr
(where pr_warning() writes):

  Error: KVM exit reason: 9 ("KVM_EXIT_FAIL_ENTRY")
  Warning: unable to translate guest address 0x0 to host

 Registers:

<snip>

*lr:
 0x00000000: <unknown>
 0x00000008: <unknown>
 0x00000010: <unknown>
 0x00000018: <unknown>

The error is caused by the VCPU migrating to a physical CPU with a
different PMU. In the example, the warning not only is unnecessary, but is
quite a distance away from the offending address and can confuse the person
looking at the code, like it happened to the patch author.

Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>
Link: https://lore.kernel.org/r/20250120161800.30270-3-alexandru.elisei@arm.com
Signed-off-by: Will Deacon <will@kernel.org>