util: Fix update_headers.sh after removal of 32-bit Arm code
Commit 4cd7b87ab292 ("arm64: Move asm headers") moved the 64-bit
Arm headers as part of the cleanup following the removal of the 32-bit
Arm code. Unfortunately, the update_headers.sh wasn't updated with the
new paths and so trying to sync the kernel UAPI results in an error
from the script:

  | cp: cannot create regular file 'arm/aarch64/include/asm/': No such file or directory

Fix-up the script so that it doesn't look in non-existent directories.

Signed-off-by: Will Deacon <will@kernel.org>